requirements.txt
# Dépendances Python pour l'agent de paris
requests>=2.31.0
pandas>=2.0.0
pyTelegramBotAPI>=4.14.0
schedule>=1.2.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
